Delay copies until mailbox close
With more recent Dovecot versions, we are not allowed to perform a
mailbox sync while there are any transactions active, and the
transaction is active throughout our overridden transaction commit
method even if we call the super's committer before doing our work.
As a workaround, save all the newly added UIDs and perform the
copies only when the mailbox is closed.